Separation of jvm-specific code from DatabaseMetaDataTest
